+++ src/main/java/bowling/domain/BowlingGame.java
@@ -1,0 +1,4 @@
+package bowling.domain;
+
+public class BowlingGame {
+}
