+++ src/main/java/bowling/domain/FrameSnapshot.java
@@ -1,6 +1,7 @@
 package bowling.domain;
 
 import bowling.domain.vo.BowlingBalls;
+import bowling.domain.vo.RoundUnit;
 
 import java.util.HashMap;
 import java.util.Map;
